Fixed bug for multiple entries with the same key
If two tv shows had episodes coming at the same day e.g.
Mr Robot S03E01 in 12 days
The Flash S0XE01 in 12 days

it would replace value based on key '12' in global variable
"HTML_OUTPUT_DICT" with the episode of tv show that was parsed next. In
this case HTML_OUTPUT_DICT[12] would replace html code of Mr. Robot with
The Flash and Mr. Robot TV show would not show up on the homepage at
all.

This was fixed using setdefault method for dictionaries, which will set
the value of a key to list [] and append items to it if two tv shows
(episodes) will  be released at the same day.